requirements.txt
Django==1.8.7
djangorestframework==3.3.1
djoser==0.4.0
wheel==0.24.0
